scoreboard, login, registration is working (no Database)
