+++ src/main/java/server/Messages/MessageStates.java
@@ -3,7 +3,7 @@ package server.Messages;
 public class MessageStates {
     public static final String ALREADY_AUTHORIZED = "user is already authorized";
     public static final String NOT_ENOUGH_DATA = "not enough data";
-    public static final String AUTHORIZED = "status:successful authorize";
+    public static final String AUTHORIZED = "successful authorize";
     public static final String BAD_AUTHORIZE = "invalid login or password";
     public static final String EMAIL_ALREADY_EXISTS = "this email already registered";
     public static final String LOGIN_ALREADY_EXISTS = "this login already registered";
